+++ WalletWasabi.Fluent/Views/AddWallet/ConfirmRecoveryWordsView.axaml.cs
@@ -1,7 +1,8 @@
 using Avalonia.Markup.Xaml;
 using Avalonia.ReactiveUI;
+using WalletWasabi.Fluent.ViewModels.AddWallet;
 
-namespace WalletWasabi.Fluent.AddWallet.CreateWallet
+namespace WalletWasabi.Fluent.Views.AddWallet
 {
 	public class ConfirmRecoveryWordsView : ReactiveUserControl<ConfirmRecoveryWordsViewModel>
 	{
