+++ WalletWasabi.Fluent/Views/AddWallet/RecoveryWordsView.axaml
@@ -7,7 +7,7 @@
              xmlns:behaviors="clr-namespace:WalletWasabi.Fluent.Behaviors;assembly=WalletWasabi.Fluent"
              xmlns:i="clr-namespace:Avalonia.Xaml.Interactivity;assembly=Avalonia.Xaml.Interactivity"
              d:DesignWidth="600" d:DesignHeight="500"
-             x:Class="WalletWasabi.Fluent.AddWallet.CreateWallet.RecoveryWordsView">
+             x:Class="WalletWasabi.Fluent.Views.AddWallet.RecoveryWordsView">
   <Panel>
 
     <c:ContentArea Title="Recovery words"
